+++ embed/src/Embed.vue
@@ -6,17 +6,29 @@
       <p v-show="embedSettings.showCoordinateReadout">{{ coordText }}</p>
     </div>
 
-    <div id="tool-menu" v-show="showToolMenu">
-      <v-popover>
-        <font-awesome-icon class="tooltip-target" icon="sliders-h" size="lg"></font-awesome-icon>
-        <template slot="popover">
-          <ul class="tooltip-content tool-menu">
-            <li v-show="showCrossfader"><a href="#" v-close-popover @click="selectTool('crossfade')"><font-awesome-icon icon="adjust" /> Crossfade</a></li>
-            <li v-show="showBackgroundChooser"><a href="#" v-close-popover @click="selectTool('choose-background')"><font-awesome-icon icon="mountain" /> Choose background</a></li>
-          </ul>
-        </template>
-      </v-popover>
-    </div>
+    <ul id="controls">
+      <li v-show="showToolMenu">
+        <v-popover placement="left">
+          <font-awesome-icon class="tooltip-target" icon="sliders-h" size="lg"></font-awesome-icon>
+          <template slot="popover">
+            <ul class="tooltip-content tool-menu">
+              <li v-show="showCrossfader"><a href="#" v-close-popover @click="selectTool('crossfade')"><font-awesome-icon icon="adjust" /> Crossfade</a></li>
+              <li v-show="showBackgroundChooser"><a href="#" v-close-popover @click="selectTool('choose-background')"><font-awesome-icon icon="mountain" /> Choose background</a></li>
+            </ul>
+          </template>
+        </v-popover>
+      </li>
+      <li v-show="!wwtIsTourPlayerActive">
+        <font-awesome-icon icon="search-plus" size="lg" @click="doZoom(true)"></font-awesome-icon>
+      </li>
+      <li v-show="!wwtIsTourPlayerActive">
+        <font-awesome-icon icon="search-minus" size="lg" @click="doZoom(false)"></font-awesome-icon>
+      </li>
+      <li v-show="fullscreenAvailable">
+        <font-awesome-icon v-bind:icon="fullscreenModeActive ? 'compress' : 'expand'"
+          size="lg" class="nudgeright1" @click="toggleFullscreen()"></font-awesome-icon>
+      </li>
+    </ul>
 
     <div id="tools">
       <div class="tool-container">
@@ -46,6 +58,8 @@
 
 <script lang="ts">
 import { Component, Prop } from "vue-property-decorator";
+
+import * as screenfull from "screenfull";
 
 import { fmtDegLat, fmtDegLon, fmtHours } from "@wwtelescope/astro";
 import { ImageSetType } from "@wwtelescope/engine-types";
@@ -82,6 +96,7 @@ export default class Embed extends WWTAwareComponent {
 
   backgroundImagesets: BackgroundImageset[] = [];
   currentTool: ToolType = null;
+  fullscreenModeActive = false;
 
   get coordText() {
     if (this.wwtRenderType == ImageSetType.sky) {
@@ -107,6 +122,10 @@ export default class Embed extends WWTAwareComponent {
 
   set foregroundOpacity(o: number) {
     this.setForegroundOpacity(o);
+  }
+
+  get fullscreenAvailable() {
+    return screenfull.isEnabled;
   }
 
   get showBackgroundChooser() {
@@ -227,11 +246,45 @@ export default class Embed extends WWTAwareComponent {
     }
   }
 
+  mounted() {
+    if (screenfull.isEnabled) {
+      screenfull.on('change', this.onFullscreenEvent);
+    }
+  }
+
+  destroyed() {
+    if (screenfull.isEnabled) {
+      screenfull.off('change', this.onFullscreenEvent);
+    }
+  }
+
   selectTool(name: ToolType) {
     if (this.currentTool == name) {
       this.currentTool = null;
     } else {
       this.currentTool = name;
+    }
+  }
+
+  doZoom(zoomIn: boolean) {
+    if (zoomIn) {
+      this.zoom(1/1.3);
+    } else {
+      this.zoom(1.3);
+    }
+  }
+
+  toggleFullscreen() {
+    if (screenfull.isEnabled) {
+      screenfull.toggle();
+    }
+  }
+
+  onFullscreenEvent() {
+    // NB: we need the isEnabled check to make TypeScript happy even though it
+    // is not necesary in practice here.
+    if (screenfull.isEnabled) {
+      this.fullscreenModeActive = screenfull.isFullscreen;
     }
   }
 }
@@ -283,14 +336,24 @@ body {
   }
 }
 
-#tool-menu {
+#controls {
   position: absolute;
   top: 0.5rem;
   right: 0.5rem;
   color: #FFF;
 
-  .tooltip-target {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+
+  li {
+    padding: 3px;
+    height: 22px;
     cursor: pointer;
+
+    .nudgeright1 {
+      padding-left: 3px;
+    }
   }
 }